backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
yt-dlp

# YouTube & Transcripts
google-api-python-client==2.116.0
youtube-transcript-api==0.6.2

# ML & Embeddings
sentence-transformers==2.3.1
faiss-cpu==1.8.0
torch==2.8.0
transformers==4.37.2

# Data Processing
pandas==2.2.0
numpy==1.26.3
pyarrow==15.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# MySQL Support
aiomysql==0.2.0
PyMySQL==1.1.0
cryptography==41.0.7

# Optional: Pinecone
pinecone-client==3.0.2

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.26.0
aiofiles==23.2.1

# Monitoring & Logging
loguru==0.7.2
